[Fix] 작성자 일치 여부 코드 수정

+++ src/main/java/com/finalproject_sns/service/PostService.java
@@ -63,7 +63,9 @@ public class PostService {
     public PostResponse update(PostRequest request, Long postId, String userName) {
         Post post = postRepository.findById(postId).orElseThrow(() -> new UserAppException(ErrorCode.POST_NOT_FOUND, "해당 글은 존재하지 않습니다."));
         User user = userRepository.findByUserName(userName).orElseThrow(() -> new UserAppException(ErrorCode.USERNAME_NOT_FOUND, userName + "은 존재하지 않는 회원입니다."));
-        if (post.getUser().getUserName() != request.toEntity(user).getUser().getUserName()) {
+
+        // 작성자가 아닐 떄
+        if (!post.getUser().getUserName().equals(request.toEntity(user).getUser().getUserName())) {
             throw new UserAppException(ErrorCode.INVALID_PERMISSION,"사용자가 권한이 없습니다.");
         }
         post.update(request.getTitle(),request.getBody());
@@ -77,10 +79,15 @@ public class PostService {
 
     public PostResponse deletePost(Long postId, String userName) {
         // user 존재하지 않을 때
-        userRepository.findByUserName(userName).orElseThrow(() -> new UserAppException(ErrorCode.USERNAME_NOT_FOUND, userName + "은 존재하지 않는 회원입니다."));
+        User user = userRepository.findByUserName(userName).orElseThrow(() -> new UserAppException(ErrorCode.USERNAME_NOT_FOUND, userName + "은 존재하지 않는 회원입니다."));
 
         // 게시글 존재하지 않을 떄
-        postRepository.findById(postId).orElseThrow(() -> new UserAppException(ErrorCode.POST_NOT_FOUND, "해당 글은 존재하지 않습니다."));
+        Post post = postRepository.findById(postId).orElseThrow(() -> new UserAppException(ErrorCode.POST_NOT_FOUND, "해당 글은 존재하지 않습니다."));
+
+        // 작성자가 아닐 떄
+        if (!post.getUser().getUserName().equals(userName)) {
+            throw new UserAppException(ErrorCode.INVALID_PERMISSION,"사용자가 권한이 없습니다.");
+        }
 
         postRepository.deleteById(postId);
         return PostResponse.builder()
